Add some more test cases and refactor
This adds a couple more test cases that were missed. This also refactors
the rule a bit to clean it up. There is also a workaround for keeping
track of TypeAlias nodes seen to work around an issue with eslint@3 and
babel-eslint preventing the TypeAlias variables being found in the
scope.